fix: keep detached asyncio tasks alive and widen the Ruff gate
asyncio holds only a weak reference to a running task, so a task that no
caller stores can be garbage collected mid-await. Six fire-and-forget
tasks were scheduled that way. The overlay stop path was the most
visible: pressing stop on the recording overlay scheduled
stop_listening() with nothing holding it, so the recording could keep
running.

Give each detached task an owner:

- ScriberWebController gains _spawn_detached()/_wait_for_detached_tasks(),
  matching the existing _metrics_persist_tasks pattern, and shutdown now
  drains them alongside the metric and transcript writes.
- MeetingLiveTranscriber holds its backpressure report task.
- _AnalyzerCache holds pending async analyzer cleanups.

Enable Ruff's ASYNC and RUF rule families so this class of defect is
caught at lint time. Rules that do not fit the codebase are switched off
with the reason recorded next to them: ASYNC109 is a Trio convention,
ASYNC230/ASYNC240 need latency measurements on meeting-length audio
before the blocking-IO sites are reworked, and RUF001 flags deliberate
typography in the localized export templates. Tests and scripts poll
external state on purpose, so ASYNC110 and the style-only RUF rules are
relaxed there via per-file-ignores; the three production polls that wait
on third-party state carry an inline noqa with its reason.

The newly enabled rules also surfaced a real typing defect:
unload_model(model_name: str = None) declared an implicit Optional.

+++ scripts/installer_research/comparator.py
@@ -560,7 +560,7 @@ def _timing_summary(
     def total_nanoseconds(installer_bytes: int, milliseconds: Any) -> int | None:
         if isinstance(milliseconds, bool) or not isinstance(milliseconds, (int, float)):
             return None
-        return installer_bytes * 160 + int(round(float(milliseconds) * 1_000_000))
+        return installer_bytes * 160 + round(float(milliseconds) * 1_000_000)
 
     return {
         "evidenceSha256": evidence_sha256,
